refactor: 🔄 split createRoundsFromCupTrees into 2 fns
